set neutral_value to p_keep
When we moved from distance weighted masks to distance ranked mask filtering we seem to have forgotten to change the neutral_value. The neutral_value was still a distance, but the saliency map values are not distances anymore, so this doesn't match. Now, we set the neutral_value to the expectation value of a random set of masks, which is equal to p_keep. Note that in fact this is may not be the expectation value of the distance rank filtered masks, but calculating a better expectation value that properly accounts for the reference image used and other parameters involved is out of scope for now.